完善全局异常返回类，全局异常状态枚举类，集成j303参数校验，初步编写注册接口

+++ train-member/src/main/java/com/train/member/mapper/MemberMapper.java
@@ -14,7 +14,7 @@ public interface MemberMapper {
 
     int insert(Member record);
 
-    int insertSelective(Member record);
+    Long insertSelective(Member record);
 
     List<Member> selectByExample(MemberExample example);
